+++ praudio/io/numpysaver.py
@@ -7,18 +7,19 @@ import logging
 
 import numpy as np
 
+from praudio.io.saver import Saver
 from praudio.utils import add_extension_to_file
 
 
 logger = logging.getLogger(__name__)
 
 
-class NumpySaver:
+class NumpySaver(Saver):
     """NumpySaver saves arrays as npy files. It's a wrapper around numpy
     saving facilities.
     """
     def __init__(self):
-        self._extension = "npy"
+        super().__init__("npy")
         logger.info("Initialised NumpySaver object")
 
     def save(self,
@@ -29,6 +30,6 @@ class NumpySaver:
         :param file: Path where to save file without extension
         :param array: Numpy array to store
         """
-        file_with_extension = add_extension_to_file(file, self._extension)
+        file_with_extension = add_extension_to_file(file, self.extension)
         np.save(file_with_extension, array)
         logger.info("Array saved at %s", file_with_extension)