+++ praudio/utils.py
@@ -2,6 +2,7 @@
 This module contains utility functions that are used throughout the library.
 """
 
+import os
 from pathlib import Path
 
 
@@ -25,3 +26,32 @@ def add_extension_to_file(file: str,
     :return: File with extension
     """
     return file + "." + extension
+
+
+def remove_extension_from_file(file: str) -> str:
+    """
+    :param file: File path to which to remove extension
+
+    :return: File without extension
+    """
+    return str(Path(file).with_suffix(""))
+
+
+def create_dir_hierarchy(dir: str):
+    """Creates a hierarchy of directories, if it doesn't exists. Else,
+    it doesn't do anything.
+
+    :param dir: Path with directory hierarchy which should be created, if it
+        doesn't exist
+    """
+    Path(dir).mkdir(parents=True, exist_ok=True)
+
+
+def create_dir_hierarchy_from_file(file: str):
+    """Creates a hierarchy of directories for a file, if it doesn't exists.
+    Else, it doesn't do anything.
+
+    :param file: File for which to create the relative dir hierarchy
+    """
+    dir = os.path.dirname(file)
+    create_dir_hierarchy(dir)